+++ phpunit.xml.dist
@@ -18,14 +18,6 @@
         </testsuite>
     </testsuites>
 
-    <!--
-    <logging>
-        <log type="coverage-html" target="build/log/report" lowUpperBound="35" highLowerBound="70" />
-        <log type="coverage-clover" target="build/log/coverage.xml" />
-        <log type="coverage-text" target="php://stdout" showUncoveredFiles="false" />
-    </logging>
-    -->
-
     <php>
         <server name="KERNEL_DIR" value="./tests/Fixtures/App" />
         <server name="KERNEL_CLASS" value="\AppKernel" />
@@ -35,25 +27,4 @@
         <ini name="memory_limit" value="-1" />
         <ini name="zend.enable_gc" value="1" />
     </php>
-
-    <filter>
-        <whitelist processUncoveredFilesFromWhitelist="true">
-            <directory suffix=".php">./src/Configuration</directory>
-            <directory suffix=".php">./src/Controller</directory>
-            <directory suffix=".php">./src/DependencyInjection</directory>
-            <directory suffix=".php">./src/Reflection</directory>
-            <directory suffix=".php">./src/Resources</directory>
-            <directory suffix=".php">./src/Twig</directory>
-            <exclude>
-                <directory>./build</directory>
-                <directory>./tests</directory>
-                <directory>./vendor</directory>
-            </exclude>
-        </whitelist>
-        <blacklist>
-            <directory>./vendor</directory>
-            <directory>./tests</directory>
-            <directory>./build</directory>
-        </blacklist>
-    </filter>
 </phpunit>
