+++ tests/Configuration/fixtures/exceptions/template_override.yml
@@ -4,7 +4,7 @@
 # EXCEPTION
 expected_exception:
     class: \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-    message_regexp: '/Unrecognized options? "this_template_name_is_not_valid" under "easy_admin.design.templates"/'
+    message_regexp: '/Unrecognized option "this_template_name_is_not_valid" under "easy_admin.design.templates"/'
 
 # CONFIGURATION
 easy_admin:
