+++ tests/Fixtures/App/config/config.yml
@@ -52,3 +52,6 @@ security:
 services:
     logger:
         class: Psr\Log\NullLogger
+
+    EasyCorp\Bundle\EasyAdminBundle\Tests\Fixtures\AppTestBundle\DataFixtures\AppFixtures:
+        tags: ['doctrine.fixture.orm']
